requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
streamlit
openpyxl

langchain_groq
langchain_core

pandas 
numpy
matplotlib
plotly-express
seaborn
openpyxl
xgboost
scikit-learn